+++ tests/foreman/ui_airgun/test_repository.py
@@ -28,7 +28,6 @@ from robottelo.constants import (
     CHECKSUM_TYPE,
     DISTRO_RHEL7,
     DOCKER_REGISTRY_HUB,
-    DOCKER_UPSTREAM_NAME,
     DOWNLOAD_POLICIES,
     FAKE_0_PUPPET_REPO,
     FAKE_1_PUPPET_REPO,
@@ -383,33 +382,6 @@ def test_positive_sync_custom_repo_docker(session, module_org):
     with session:
         result = session.repository.synchronize(product.name, repo.name)
         assert result['result'] == 'success'
-
-
-@tier2
-@upgrade
-def test_positive_sync_docker_via_sync_status(session, module_org):
-    """Create custom docker repo and sync it via the sync status page.
-
-    :id: 00b700f4-7e52-48ed-98b2-e49b3be102f2
-
-    :expectedresults: Sync procedure for specific docker repository is
-        successful
-
-    :CaseLevel: Integration
-    """
-    product = entities.Product(organization=module_org).create()
-    repo_name = gen_string('alphanumeric')
-    with session:
-        session.repository.create(
-            product.name,
-            {'name': repo_name,
-             'repo_type': REPO_TYPE['docker'],
-             'repo_content.upstream_url': DOCKER_REGISTRY_HUB,
-             'repo_content.upstream_repo_name': DOCKER_UPSTREAM_NAME}
-        )
-        assert session.repository.search(product.name, repo_name)[0]['Name'] == repo_name
-        result = session.sync_status.synchronize([(product.name, repo_name)])
-        assert result[0] == 'Syncing Complete.'
 
 
 @tier2
